*** empty log message ***

+++ src/api/mysql/mysql.cpp
@@ -1908,7 +1908,7 @@ static void getDate(const char *field, uint32_t length, MYSQL_BIND *bind) {
 	int16_t	hour=-1;
 	int16_t	minute=-1;
 	int16_t	second=-1;
-	int16_t	fraction=-1;
+	int32_t	fraction=-1;
 	bool	isnegative=false;
 
 	// copy into a buffer (to make sure it's null-terminated)
