+++ src/common/parsedatetime.h
@@ -50,7 +50,7 @@ static bool parseDateTime(const char *datetime, bool ddmm, bool yyyyddmm,
 			const char *datedelimiters,
 			int16_t *year, int16_t *month, int16_t *day,
 			int16_t *hour, int16_t *minute, int16_t *second,
-			int16_t *fraction, bool *isnegative) {
+			int32_t *fraction, bool *isnegative) {
 
 	bool	supportslashdelimiteddate=
 			charstring::contains(datedelimiters,'/');
@@ -640,7 +640,7 @@ static bool parseDateTime(const char *datetime, bool ddmm, bool yyyyddmm,
 static char *convertDateTime(const char *format,
 			int16_t year, int16_t month, int16_t day,
 			int16_t hour, int16_t minute, int16_t second,
-			int16_t fraction, bool isnegative) {
+			int32_t fraction, bool isnegative) {
 
 	// if no format was passed in
 	if (!format) {
